move logger impl to connector-service and clean up pom
