--- java/connector-node/connector-api/src/main/resources/META-INF/services/org.slf4j.spi.SLF4JServiceProvider
@@ -1,1 +1,0 @@
-com.risingwave.connector.api.tracing.TracingSlf4jServiceProvider
