+++ java/connector-node/risingwave-connector-service/src/main/java/com/risingwave/tracing/TracingSlf4jAdapter.java
@@ -15,7 +15,7 @@
 // Ported from https://github.com/MrFriendly-B-V/tracing-slf4j,
 // which is licensed under the Apache License, Version 2.0.
 
-package com.risingwave.connector.api.tracing;
+package com.risingwave.tracing;
 
 import org.slf4j.Logger;
 import org.slf4j.Marker;
