+++ java/connector-node/risingwave-connector-service/src/main/java/com/risingwave/tracing/TracingSlf4jImpl.java
@@ -15,14 +15,18 @@
 // Ported from https://github.com/MrFriendly-B-V/tracing-slf4j,
 // which is licensed under the Apache License, Version 2.0.
 
-package com.risingwave.connector.api.tracing;
+package com.risingwave.tracing;
 
-import org.slf4j.ILoggerFactory;
-import org.slf4j.Logger;
+import com.risingwave.java.binding.Binding;
 
-public class TracingSlf4jLoggerFactory implements ILoggerFactory {
-    @Override
-    public Logger getLogger(String name) {
-        return new TracingSlf4jAdapter(name);
+public class TracingSlf4jImpl {
+    public static final int ERROR = 0;
+    public static final int WARN = 1;
+    public static final int INFO = 2;
+    public static final int DEBUG = 3;
+    public static final int TRACE = 4;
+
+    public static void event(String name, int level, String message) {
+        Binding.tracingSlf4jEvent(name, level, message);
     }
 }
