+++ java/connector-node/risingwave-connector-service/src/main/resources/META-INF/services/org.slf4j.spi.SLF4JServiceProvider
@@ -1,0 +1,1 @@
+com.risingwave.tracing.TracingSlf4jServiceProvider
