+++ java/connector-node/risingwave-connector-test/pom.xml
@@ -38,16 +38,6 @@
             <groupId>com.risingwave</groupId>
             <artifactId>s3-common</artifactId>
             <scope>test</scope>
-            <exclusions>
-                <exclusion>
-                    <groupId>org.slf4j</groupId>
-                    <artifactId>slf4j-log4j12</artifactId>
-                </exclusion>
-                <exclusion>
-                    <groupId>org.slf4j</groupId>
-                    <artifactId>slf4j-reload4j</artifactId>
-                </exclusion>
-            </exclusions>
         </dependency>
         <dependency>
             <groupId>io.delta</groupId>
