+++ java/connector-node/risingwave-sink-iceberg/pom.xml
@@ -75,18 +75,6 @@
             <artifactId>url-connection-client</artifactId>
             <version>2.18.20</version>
         </dependency>
-        <!-- <dependency>
-            <groupId>org.apache.logging.log4j</groupId>
-            <artifactId>log4j-api</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.logging.log4j</groupId>
-            <artifactId>log4j-slf4j-impl</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.logging.log4j</groupId>
-            <artifactId>log4j-core</artifactId>
-        </dependency> -->
 
     </dependencies>
 </project>
